Move generic mobile API view decorators.
