+++ lms/djangoapps/mobile_api/course_info/views.py
@@ -87,33 +87,3 @@ class CourseHandoutsList(generics.ListAPIView):
         else:
             # course_handouts_module could be None if there are no handouts
             raise Http404(u"No handouts for {}".format(unicode(course.id)))
-
-
-@mobile_view()
-class CourseAboutDetail(generics.RetrieveAPIView):
-    """
-    **Use Case**
-
-        Get the HTML for the course about page.
-
-    **Example request**:
-
-        GET /api/mobile/v0.5/course_info/{organization}/{course_number}/{course_run}/about
-
-    **Response Values**
-
-        * overview: The HTML for the course About page.
-    """
-
-    @mobile_course_access(verify_enrolled=False)
-    def get(self, request, course, *args, **kwargs):
-        # There are other fields, but they don't seem to be in use.
-        # see courses.py:get_course_about_section.
-        #
-        # This can also return None, so check for that before calling strip()
-        about_section_html = get_course_about_section(course, "overview")
-        about_section_html = make_static_urls_absolute(self.request, about_section_html)
-
-        return Response(
-            {"overview": about_section_html.strip() if about_section_html else ""}
-        )
