+++ lms/djangoapps/mobile_api/users/views.py
@@ -212,7 +212,6 @@ class UserCourseEnrollmentsList(generics.ListAPIView):
           * url: URL to the downloadable version of the certificate, if exists.
         * course: A collection of data about the course:
 
-          * course_about: The URI to get the data for the course About page.
           * course_updates: The URI to get data for course updates.
           * number: The course number.
           * org: The organization that created the course.
