+++ openedx/core/lib/api/permissions.py
@@ -1,6 +1,5 @@
 from django.conf import settings
 from rest_framework import permissions
-from rest_framework.exceptions import PermissionDenied
 from django.http import Http404
