+++ openedx/core/lib/api/view_utils.py
@@ -1,11 +1,24 @@
 """
 Utilities related to API views
 """
+import functools
 from django.core.exceptions import NON_FIELD_ERRORS, ValidationError
 from django.http import Http404
 
+from rest_framework import status, response
 from rest_framework.exceptions import APIException
 from rest_framework.response import Response
+
+from lms.djangoapps.courseware.courses import get_course_with_access
+from opaque_keys.edx.keys import CourseKey
+from xmodule.modulestore.django import modulestore
+
+from openedx.core.lib.api.authentication import (
+    SessionAuthenticationAllowInactiveUser,
+    OAuth2AuthenticationAllowInactiveUser,
+)
+from openedx.core.lib.api.permissions import IsUserInUrl, IsAuthenticatedOrDebug
+from util.milestones_helpers import any_unfulfilled_milestones
 
 
 class DeveloperErrorViewMixin(object):
@@ -48,3 +61,60 @@ class DeveloperErrorViewMixin(object):
             return self.make_validation_error_response(exc)
         else:
             raise
+
+
+def view_course_access(depth=0, access_action='load', check_for_milestones=False):
+    """
+    Method decorator for an API endpoint that verifies the user has access to the course.
+    """
+    def _decorator(func):
+        """Outer method decorator."""
+        @functools.wraps(func)
+        def _wrapper(self, request, *args, **kwargs):
+            """
+            Expects kwargs to contain 'course_id'.
+            Passes the course descriptor to the given decorated function.
+            Raises 404 if access to course is disallowed.
+            """
+            course_id = CourseKey.from_string(kwargs.pop('course_id'))
+            with modulestore().bulk_operations(course_id):
+                try:
+                    course = get_course_with_access(
+                        request.user,
+                        access_action,
+                        course_id,
+                        depth=depth
+                    )
+                except Http404:
+                    # any_unfulfilled_milestones called a second time since has_access returns a bool
+                    if check_for_milestones and any_unfulfilled_milestones(course_id, request.user.id):
+                        message = {
+                            "developer_message": "Cannot access content with unfulfilled "
+                                                 "pre-requisites or unpassed entrance exam."
+                        }
+                        return response.Response(data=message, status=status.HTTP_204_NO_CONTENT)
+                    else:
+                        raise
+                return func(self, request, course=course, *args, **kwargs)
+        return _wrapper
+    return _decorator
+
+
+def view_auth_classes(is_user=False):
+    """
+    Function and class decorator that abstracts the authentication and permission checks for api views.
+    """
+    def _decorator(func_or_class):
+        """
+        Requires either OAuth2 or Session-based authentication.
+        If is_user is True, also requires username in URL matches the request user.
+        """
+        func_or_class.authentication_classes = (
+            OAuth2AuthenticationAllowInactiveUser,
+            SessionAuthenticationAllowInactiveUser
+        )
+        func_or_class.permission_classes = (IsAuthenticatedOrDebug,)
+        if is_user:
+            func_or_class.permission_classes += (IsUserInUrl,)
+        return func_or_class
+    return _decorator
